// src-ui/core/ipc/generated/SnowLumaAgreementsPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SnowLumaAgreementDoc } from "./SnowLumaAgreementDoc";

export type SnowLumaAgreementsPayload = { version: string, consent_required: boolean, documents: Array<SnowLumaAgreementDoc>, };