// src-ui/core/ipc/generated/ConfigDrift.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BackendType } from "./domain/BackendType";
import type { DriftEntry } from "./DriftEntry";

export type ConfigDrift = { bot_id: string, backend_type: BackendType, added: Array<DriftEntry>, modified: Array<DriftEntry>, };